This commit contains the implementation of the new vegetation heat capacity function.  It also includes the implementation of various instances of vegetation_energy splitting and combining when it is needed during disturbance, leaf drop, leaf out, fusion, fission, etc.  Basically, we need to make sure we always update veg_energy whenever the cohort changes its 1)number density, 2) leaf biomass, 3) dead biomass or 4) vegetation water. In many cases where splitting and combining of cohorts occurs, it was chosen to re-calculate vegetation energy based on the cohort temperature which should be uneffected by some of these processes.
The current calculation of vegetation heat capacity uses a linear combination of leaf mass times 
the pft-dependant specific heat + dead mass times a factor times the pft-dependant specific heat
for dead biomass.  The factor is the percentage of dead biomass we want to consider a "stem"
that is part of the energy balance.  Right now it is set to 5%, but it can be set to 
zero or any other number we so choose.

Also, this method of using absolute cohort biomass for calculation of heat capacity
is not without its instabilities.  The first time I ran the code without using 
a minimum heat capacity threshold, the model crashed pretty quickly.  You will see in the
heat capacity function, that the mimimum threshold somewhat mimicks the value that
we were getting with the old scheme.  It is a pretty conservative (high) threshold
right now, and maybe we can slowly lower it and get stable results.