Added functionality to do ied_init_mode=42, allowing for the INITIAL type land-surface initialization procedure that copies ed2.1 type vegetation structure and composition data along with soil texture and soil moisture initial condition.
This new functionality introduces a new alternative subroutine 
read_ed21_polyclone in ed_read_ed21_history.F90.  This is where the information is read in from the files.  This routine is based off of read_ed21_history_unstruct.

Modifications were made to opspec to allow for mode 42 on the ED2IN read, ed_init.f90 so that read_ed21_polyclone is called for the correct condition.  The subroutine allocate_polygon (site level vectors) needs to be avoided prior to read_ed21_polyclone, and this bypass is applied to ed_init.f90.  The last modification places an if condition in ed_init_atm.f90 so that soil moisture is NOT overwritten by the scalar parameter in this new case.

This release does not have extensive testing, nor are the RAMSIN or ED2IN files changed, those will be added in the next commit.