Possible fix to 'explicit interface' error
Without the `-fno-whole-file` flag, gfortran can't call subroutines in
other subroutines unless they are defined as modules. This implements a
crude fix where required "missing" subroutines are moved into a module
at the top of the relevant files.

This solves the 'explicit interface' error, but still fails to compile,
though that may be related to other issues.